Add update_menu_item api

+++ lib/plate_slate_web/schema/menu_types.ex
@@ -18,7 +18,7 @@ defmodule PlateSlateWeb.Schema.MenuTypes do
     field :priced_above, :float
 
     @desc "Priced below a value"
-    field :priced_below, :float
+    field :priced_below, :decimal
 
     @desc "Added to the menu before this date"
     field :added_before, :date
